Mark the mojo as thread safe for parallel builds
It seems safe to mark this plugin as thread safe for parallel builds
(https://cwiki.apache.org/MAVEN/parallel-builds-in-maven-3.html). Using
the threadSafe annotation requires upgrading maven-plugin-plugin. The
other change in this commit (converting @parameter to @component) is
just to get rid of a warning message that the upgraded
maven-plugin-plugin displays during the build of proguard-maven-plugin
itself.